+++ plugins/myzsh/myzsh.plugin.zsh
@@ -139,8 +139,27 @@ function handle_diffs() {
 		fi 
 }
 
+# HANDLE .ZSHRC IF/WHEN NECESSARY (if ~/.vimrc exists and $CHECK_MYTMUX exists; -z is true if it does not exist. that's why the ! at the beginning to reverse make it true if it doesn't not exist). 
+# The reason for these checks is because:
+# 1. The plugins describe how to implement what's inside those plugins. It is natural for "your" various RC files to be custom. 
+# 2. Based on #1 it is natural to want to ensure that your system is running the expected configuration. 
+# 3. There are sometimes conflicts between what is "expected" when you go through the natural progression of trying new things. 
+# a. You could change your ~/.mydotfiles/plugins/<custom-plugin>/.rcfile and then just always rely on the custom-plugin enforcer to always write that changed/updated file to its normal location.
+# b. I prefer to make changes, not to my <custom-plugin>/.rcfile, but instead to the normally implemented .rcfile. I like this because:
+# I. If I don't like the changes then I can just use my last approved/expected configuration from my <custom-plugin>. I don't need to sit there removing configurations, etc..
+# II. If I do like the change then I copy over those changes into my <custom-plugin>/.rcfile and then push that change to git. 
+# However, this lifecycle comes with challenges. Namely: Instead of always writing <custom-plugin>/.rc file to its normal location we need to detect when there are differences and then how to handle those differences. 
+# Yet, handling changes comes with its own challenges. If you're using a shared repository of <custom-plugin>'s then you may have users that are impacted by the steps taken to handle the lifecycle. 
+# For example, if a user uses a plugin that has this lifecycle handling AND the user is also used to sudo, su, create new shells, etc.. then these new shells could be negatively impacted (hang) by a prompt for how to handle the changes.
+# This problem is handled by using a unique CHECK environment variable. If it is available then check the file to determine appropriate course of action. 
+# Caveats:
+# a. Adding/Removing the CHECK variables are done through the .zshrc file. It was designed this way so that the same plugins could be used by many (app users). It isn't defined in the plugin since you'd then need two separate plugins (since you don't want a git pull to constantly overwrite this customization). So the idea is that app users will likely be created by something like Ansible which you simply add a specific step for the users where you want this enabled to update your preference of tracking changes. 
+# b. 
 
-if [ -f ${homefile} ]; then {
+# This next line (export ...) should be commented out if you do not want to manage this file and get prompted on how to handle changes between this plugin's managedFile and the file used at runtime by the program.  
+# For more information see "Feture: Manging files" in the README.md. 
+export CHECK_MYZSHRC=anything
+if [ -f ${homefile} ] && [ ! -z ${CHECK_MYZSHRC} ]; then {
   if ! cmp -s ${pluginfile} ${homefile}; then {
     warn "Differences detected: ${pluginfile} ${homefile}"
 
@@ -167,7 +186,16 @@ if [ -f ${homefile} ]; then {
 	}
 	fi
 } else {
-	# $HOME/.zshrc didn't exist. Copying plugin .vimrc file to ${HOME}
-  cp ${pluginfile} ${homefile} 
+	if [ -f ${homefile} ]; then {
+		backup_file_suffix=$(date +%Y-%m-%d_%H:%M:%S | sed 's/\(\.[0-9][0-9][0-9]\)[0-9]*$/\1/' | sed -E 's/((\s)|(:))/./g')
+		info "Backing up ${homefile}  to  ${homefile}.bak.${backup_file_suffix}"
+		cp ${homefile} ${homefile}.bak.${backup_file_suffix}
+		info "Replacing ${homefile}  with  ${pluginfile}"
+		cp ${pluginfile} ${homefile}
+	} else {
+		# $HOME/.zshrc didn't exist. Copying plugin .zshrc file to ${HOME}
+  	cp ${pluginfile} ${homefile} 
+	}
+	fi
 }
 fi
